refactor(step2):  상속구조 변경
extends 지양, enum안에서 인터페이스로 처리
