--- src/main/java/calculator/operator/MinusOperator.java
@@ -1,9 +1,0 @@
-package calculator.operator;
-
-public class MinusOperator extends Operator {
-
-    @Override
-    public double calculate(double operand1, double operand2) {
-        return operand1 - operand2;
-    }
-}
