--- src/main/java/calculator/operator/Operator.java
@@ -1,6 +1,0 @@
-package calculator.operator;
-
-public abstract class Operator {
-
-    abstract public double calculate(double operand1, double operand2);
-}
