+++ src/main/java/calculator/operator/OperatorEnum.java
@@ -2,31 +2,45 @@ package calculator.operator;
 
 public enum OperatorEnum {
 
-    PLUS("+", new PlusOperator()),
-    MINUS("-", new MinusOperator()),
-    MULTIPLY("*", new MultiplyOperator()),
-    DIVISION("/", new DivisionOperator());
+    PLUS("+") {
+        public double calculate(double operand1, double operand2) {
+            return operand1 + operand2;
+        }
+    },
+    MINUS("-") {
+        public double calculate(double operand1, double operand2) {
+            return operand1 - operand2;
+        }
+    },
+    MULTIPLY("*") {
+        public double calculate(double operand1, double operand2) {
+            double result = operand1 * operand2;
+            return Math.round(result * 100) / 100.0;
+        }
+    },
+    DIVISION("/") {
+        public double calculate(double operand1, double operand2) {
+            double result = operand1 / operand2;
+            return Math.round(result * 100) / 100.0;
+        }
+    };
+
+    public abstract double calculate(double operand1, double operand2);
 
     private String keyword;
-    private Operator operator;
 
-    OperatorEnum(String keyword, Operator operator) {
+    OperatorEnum(String keyword) {
         this.keyword = keyword;
-        this.operator = operator;
     }
 
     public String getKeyword() {
         return keyword;
     }
 
-    public Operator getOperator() {
-        return operator;
-    }
-
-    public static Operator getOperator(String operatorKeyword) {
+    public static OperatorEnum getOperatorEnumFromKeyword(String operatorKeyword) {
         for (OperatorEnum operator : OperatorEnum.values()) {
             if (operator.getKeyword().equals(operatorKeyword)) {
-                return operator.getOperator();
+                return operator;
             }
         }
         throw new IllegalArgumentException("수식 중 사칙연산 기호가 잘못되었습니다.");
